Support save state and resume

+++ cmd/main.go
@@ -2,8 +2,13 @@ package main
 
 import (
 	"fmt"
+	"io/ioutil"
+	"log"
 	"os"
+	"os/signal"
+	"path/filepath"
 	"runtime"
+	"syscall"
 
 	"github.com/spf13/cobra"
 )
@@ -16,7 +21,7 @@ func main() {
 		Short: "Download the given url",
 		Args:  cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
-			get(args[0], conn)
+			get(args[0], conn, nil)
 		},
 	}
 
@@ -25,7 +30,7 @@ func main() {
 		Short: "Resume an unfinished task",
 		Args:  cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Println("TODO: Implement resume task")
+			resume(args[0], conn)
 		},
 	}
 
@@ -34,7 +39,7 @@ func main() {
 		Short: "Listing all unfinished tasks",
 		Args:  cobra.MinimumNArgs(0),
 		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Println("TODO: Implement task listing")
+			listTasks()
 		},
 	}
 
@@ -43,17 +48,33 @@ func main() {
 	rootCmd.Execute()
 }
 
-func get(url string, conn int) {
+func get(url string, conn int, state *State) {
 	var err error
-	var fileChan = make(chan string, int64(conn))
-	var doneChan = make(chan bool, int64(conn))
-	var errorChan = make(chan error, 1)
-	var files = make([]string, 0)
-	var filename = FilenameFromURL(url)
+	// var state State
+	fileChan := make(chan string, int64(conn))
+	doneChan := make(chan bool, int64(conn))
+	errorChan := make(chan error, 1)
 
-	downloader := NewHttpDownloader(url, int64(conn))
+	files := make([]string, 0)
+	parts := make([]Part, 0)
+	filename := FilenameFromURL(url)
+	signalChan := make(chan os.Signal, 1)
+	stateChan := make(chan Part, 1)
+	interupted := false
+	signal.Notify(signalChan,
+		syscall.SIGHUP,
+		syscall.SIGINT,
+		syscall.SIGTERM,
+		syscall.SIGQUIT)
 
-	go downloader.Start(doneChan, fileChan, errorChan)
+	var downloader *HttpDownloader
+	if state == nil {
+		downloader = NewHttpDownloader(url, int64(conn), make([]Part, 0))
+	} else {
+		downloader = NewHttpDownloader(url, int64(conn), state.Parts)
+	}
+
+	go downloader.Start(doneChan, fileChan, errorChan, signalChan, stateChan)
 
 	for {
 		select {
@@ -61,12 +82,44 @@ func get(url string, conn int) {
 			files = append(files, file)
 		case err := <-errorChan:
 			HandleError(err)
+		case part := <-stateChan:
+			parts = append(parts, part)
+			interupted = true
 		case <-doneChan:
+			if interupted && downloader.resumable {
+				fmt.Printf("Interrupted, saving state ... \n")
+				s := &State{URL: url, Parts: parts}
+				err = s.Save()
+				HandleError(err)
+				return
+			}
 			err = JoinFile(files, filename)
 			HandleError(err)
 			err = os.RemoveAll(GetValidFolderPath(url))
 			HandleError(err)
 			return
 		}
+	}
+}
+
+func resume(task string, conn int) {
+	state, err := LoadState(task)
+	HandleError(err)
+	get(state.URL, conn, state)
+}
+
+func listTasks() {
+	files, err := ioutil.ReadDir(filepath.Join(GetUserHome(), appHome))
+
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	if len(files) == 0 {
+		fmt.Println("There is no unfinished task")
+	}
+
+	for _, f := range files {
+		fmt.Println(f.Name())
 	}
 }
